% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aucAk}
\alias{aucAk}
\title{compute auc}
\usage{
aucAk(response, predictor)
}
\arguments{
\item{response}{input boolean vector}

\item{predictor}{input numerical vector}
}
\value{
auc
}
\description{
compute auc
}

